新しい人物追加用イベントのMVP化 ・DialogFragmentの別クラス化 ・Fragmentクラスに記載していたRealmDB操作をRepositoryに移設
